Create a CaseDetailsRow folder & move SingleException component in there
